+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/servers/server/secondary-dns/add/index.js
@@ -1,0 +1,22 @@
+import angular from 'angular';
+
+import '@uirouter/angularjs';
+import 'oclazyload';
+
+const moduleName = 'ovhManagerDedicatedServerSecondaryDnsAddLazyLoading';
+
+angular.module(moduleName, ['ui.router', 'oc.lazyLoad']).config(
+  /* @ngInject */ ($stateProvider) => {
+    $stateProvider.state('app.dedicated-server.server.secondary-dns.add.**', {
+      url: '/add',
+      lazyLoad: ($transition$) => {
+        const $ocLazyLoad = $transition$.injector().get('$ocLazyLoad');
+        return import('./secondary-dns-add.module').then((mod) =>
+          $ocLazyLoad.inject(mod.default || mod),
+        );
+      },
+    });
+  },
+);
+
+export default moduleName;