+++ packages/manager/apps/dedicated/client/app/dedicated/server/monitoring/update/dedicated-server-monitoring-update.constants.js
@@ -1,6 +1,0 @@
-export const MONITORING_TRACKING_PREFIX =
-  'dedicated::dedicated-server::server::dashboard::monitoringUpdate';
-
-export default {
-  MONITORING_TRACKING_PREFIX,
-};