+++ packages/manager/apps/web/client/app/dns-zone/history/dashboard/dashboard.component.js
@@ -1,0 +1,13 @@
+import controller from './dashboard.controller';
+import template from './dashboard.html';
+
+export default {
+  controller,
+  template,
+  bindings: {
+    goBack: '<',
+    goToDiffViewer: '<',
+    goToDnsRestore: '<',
+    goToDnsData: '<',
+  },
+};