+++ packages/manager/core/generator/utils/api-template.js
@@ -1,17 +1,18 @@
 /* eslint-disable import/extensions */
 import {
   isArrayType,
+  cleanTypeSyntax,
   getTypeFromString,
   isUnknownTypescriptType,
   removeTypeBrackets,
   transformTypeToTypescript,
 } from './string-helpers.js';
-import { getApiServiceOperations } from './api.js';
+import { getApiServiceOperations, isV2Endpoint } from './api.js';
 
 /**
  * Tranform endpoint data into a list of operations template data
  */
-const toOperationList = ({ path, description, operations }) =>
+const toOperationList = (apiVersion) => ({ path, description, operations }) =>
   operations.map(
     ({
       httpMethod,
@@ -21,6 +22,8 @@ const toOperationList = ({ path, description, operations }) =>
     }) => {
       const functionName = `${httpMethod.toLowerCase()}${path
         .replace('Name}', '}')
+        .replace('-', '')
+        .replace('/', '')
         .replace(
           /{([a-z])([a-zA-Z]+)}/g,
           (_, firstLetter, rest) => firstLetter.toUpperCase() + rest,
@@ -33,20 +36,24 @@ const toOperationList = ({ path, description, operations }) =>
 
       return {
         functionName,
+        apiVersion,
         apiPath: path,
         description: `${description} : ${operationDescription}`,
         httpMethod: httpMethod.toLowerCase(),
-        responseType: tsResponseType,
+        responseType: tsResponseType.replaceAll('>', ''),
         params: parameters.map(
           ({ name, fullType, required, description: paramDescription }) => {
-            const paramType = getTypeFromString(fullType);
-            const paramName =
+            const paramType = fullType
+              ? getTypeFromString(fullType)
+              : 'unknown';
+            let paramName =
               name ||
               `${paramType[0].toLowerCase()}${paramType
                 .substring(1)
-                .toLowerCase()}`;
+                .toLowerCase()}Custom`;
+            let nameCustom = paramName;
             return {
-              name: paramName,
+              name: nameCustom.replaceAll('.', ''),
               type: transformTypeToTypescript(paramType),
               required,
               description: paramDescription,
@@ -59,7 +66,8 @@ const toOperationList = ({ path, description, operations }) =>
             ...parameters.map(({ fullType }) => getTypeFromString(fullType)),
           ]
             .filter(isUnknownTypescriptType)
-            .map(removeTypeBrackets),
+            .map(removeTypeBrackets)
+            .map(cleanTypeSyntax),
         ),
       };
     },
@@ -123,13 +131,16 @@ const groupOperationsByHttpMethod = (groupedByMethod, operation) => {
 /**
  * @returns template data for API service operations
  */
-export const getApiv6TemplateData = async (apiPath) => {
-  const endpoints = await getApiServiceOperations(apiPath);
-  const result = endpoints
-    .flatMap(toOperationList)
+export const getApiTemplateData = async (apiPaths) => {
+  let endpoints = [];
+  for (const path of apiPaths) {
+    const result = await getApiServiceOperations(path);
+    endpoints = endpoints.concat(result);
+  }
+  return endpoints
+    .flatMap(toOperationList(isV2Endpoint(apiPaths[0]) ? 'v2' : 'v6'))
     .map(setUrlAndBodyParams)
     .reduce(groupOperationsByHttpMethod, {});
-  return result;
 };
 
-export default getApiv6TemplateData;
+export default getApiTemplateData;