+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/backups/backups.controller.js
@@ -22,10 +22,11 @@ export default class PciProjectStorageDatabaseBackupsCtrl {
     this.backupTime = this.database.backupTime;
     const flavor = this.getCurrentFlavor();
     this.backupRetentionTime = moment.duration(
-      `P${flavor.availability[0].backupRetentionDays}D`,
+      `P${flavor.availabilities[0].backups.retentionDays}D`,
     );
 
-    this.isEngineVersionDeprecated = flavor.isDeprecated;
+    this.isEngineVersionDeprecated =
+      flavor.isDeprecated || this.database.isOldMongoPlan;
     this.isPITRActivated = this.isFeatureActivated(
       'forkPIT',
       this.database.engine,