+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/general-information/general-information.controller.js
@@ -6,6 +6,7 @@ import {
   SSL_MODE_REQUIRED,
   SSL_MODE_NA,
   SSL_MODE_SSL_TLS,
+  NODES_PER_ROW,
 } from '../../databases.constants';
 import {
   WARNING_MESSAGES,
@@ -13,7 +14,10 @@ import {
   TRANSLATION_PREFIX,
 } from './general-information.constants';
 import isFeatureActivated from '../../features.constants';
-import { DISK_TYPE } from '../../../../../../components/project/storages/databases/databases.constants';
+import {
+  DISK_TYPE,
+  PLANS_WITHOUT_BACKUP,
+} from '../../../../../../components/project/storages/databases/databases.constants';
 
 export default class {
   /* @ngInject */
@@ -32,7 +36,9 @@ export default class {
     this.ovhManagerRegionService = ovhManagerRegionService;
     this.DatabaseService = DatabaseService;
     this.DATABASE_TYPES = DATABASE_TYPES;
+    this.NODES_PER_ROW = NODES_PER_ROW;
     this.user = coreConfig.getUser();
+    this.PLANS_WITHOUT_BACKUP = PLANS_WITHOUT_BACKUP;
   }
 
   $onInit() {
@@ -78,6 +84,7 @@ export default class {
       this.database.engine === DATABASE_TYPES.MONGO_DB &&
       this.database.plan === 'enterprise';
     this.isForkButtonEnabled =
+      !this.database.isOldMongoPlan &&
       !isEngineVersionDeprecated &&
       (isPITRActivated || isMongoEnterprise || this.backups.length > 0);
   }
@@ -423,6 +430,5 @@ export default class {
 
   $onDestroy() {
     this.stopPollingDatabaseStatus();
-    this.stopPollingNodesStatus();
   }
 }