+++ docs/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-documentation@0.3.0...@ovh-ux/manager-documentation@0.3.1) (2023-12-05)
+
+
+### Bug Fixes
+
+* add super-component workspaces ([a1c35e6](https://github.com/ovh/manager/commit/a1c35e6817d3f41925954c16b381ebeaea440bd7))
+
+
+
+
+
 # [0.3.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-documentation@0.2.2...@ovh-ux/manager-documentation@0.3.0) (2023-11-21)