+++ packages/manager/apps/exchange/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.15](https://github.com/ovh/manager/compare/@ovh-ux/manager-exchange-app@0.8.14...@ovh-ux/manager-exchange-app@0.8.15) (2023-12-11)
+
+**Note:** Version bump only for package @ovh-ux/manager-exchange-app
+
+
+
+
+
+## [0.8.14](https://github.com/ovh/manager/compare/@ovh-ux/manager-exchange-app@0.8.13...@ovh-ux/manager-exchange-app@0.8.14) (2023-12-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-exchange-app
+
+
+
+
+
+## [0.8.13](https://github.com/ovh/manager/compare/@ovh-ux/manager-exchange-app@0.8.12...@ovh-ux/manager-exchange-app@0.8.13) (2023-12-05)
+
+**Note:** Version bump only for package @ovh-ux/manager-exchange-app
+
+
+
+
+
 ## [0.8.12](https://github.com/ovh/manager/compare/@ovh-ux/manager-exchange-app@0.8.11...@ovh-ux/manager-exchange-app@0.8.12) (2023-12-01)
 
 **Note:** Version bump only for package @ovh-ux/manager-exchange-app