+++ packages/manager/apps/nasha/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.15](https://github.com/ovh/manager/compare/@ovh-ux/manager-nasha-app@3.0.14...@ovh-ux/manager-nasha-app@3.0.15) (2023-12-11)
+
+**Note:** Version bump only for package @ovh-ux/manager-nasha-app
+
+
+
+
+
+## [3.0.14](https://github.com/ovh/manager/compare/@ovh-ux/manager-nasha-app@3.0.13...@ovh-ux/manager-nasha-app@3.0.14) (2023-12-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-nasha-app
+
+
+
+
+
+## [3.0.13](https://github.com/ovh/manager/compare/@ovh-ux/manager-nasha-app@3.0.12...@ovh-ux/manager-nasha-app@3.0.13) (2023-12-05)
+
+**Note:** Version bump only for package @ovh-ux/manager-nasha-app
+
+
+
+
+
 ## [3.0.12](https://github.com/ovh/manager/compare/@ovh-ux/manager-nasha-app@3.0.11...@ovh-ux/manager-nasha-app@3.0.12) (2023-12-01)
 
 **Note:** Version bump only for package @ovh-ux/manager-nasha-app