+++ packages/manager/apps/netapp/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.16](https://github.com/ovh/manager/compare/@ovh-ux/manager-netapp-app@0.6.15...@ovh-ux/manager-netapp-app@0.6.16) (2023-12-11)
+
+**Note:** Version bump only for package @ovh-ux/manager-netapp-app
+
+
+
+
+
+## [0.6.15](https://github.com/ovh/manager/compare/@ovh-ux/manager-netapp-app@0.6.14...@ovh-ux/manager-netapp-app@0.6.15) (2023-12-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-netapp-app
+
+
+
+
+
+## [0.6.14](https://github.com/ovh/manager/compare/@ovh-ux/manager-netapp-app@0.6.13...@ovh-ux/manager-netapp-app@0.6.14) (2023-12-05)
+
+**Note:** Version bump only for package @ovh-ux/manager-netapp-app
+
+
+
+
+
 ## [0.6.13](https://github.com/ovh/manager/compare/@ovh-ux/manager-netapp-app@0.6.12...@ovh-ux/manager-netapp-app@0.6.13) (2023-12-01)
 
 **Note:** Version bump only for package @ovh-ux/manager-netapp-app