+++ packages/manager/apps/web/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [18.9.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@18.8.2...@ovh-ux/manager-web@18.9.0) (2023-12-11)
+
+
+### Bug Fixes
+
+* **i18n:** add missing translations [CDS 2650] ([#10462](https://github.com/ovh/manager/issues/10462)) ([c532c3b](https://github.com/ovh/manager/commit/c532c3b65e27edf84a31dd5de02076ecdf64d306))
+* **web.databases:** add missing param for shared SQL database copy ([a4dae26](https://github.com/ovh/manager/commit/a4dae264e16779968c283bf95d6912ead09ff7cb))
+
+
+### Features
+
+* **web-cloud:** starter offer - attach subdomain checkbox disabled ([#10392](https://github.com/ovh/manager/issues/10392)) ([1428576](https://github.com/ovh/manager/commit/1428576db07cde41d4a724ebb229d9ec47e5a682))
+* **web.domain:** modify the information message about free hosting ([#10355](https://github.com/ovh/manager/issues/10355)) ([23bfcc9](https://github.com/ovh/manager/commit/23bfcc9e9f3ed41ae2484fc6caf51646c6c04da5))
+* **web.hosting:** backup dbs - add filters and sort by snapshot date ([8385163](https://github.com/ovh/manager/commit/8385163f06df96901e6116e4959dc7f57162a047))
+* **web:** add additionnal message and redirect ([#10270](https://github.com/ovh/manager/issues/10270)) ([a54a891](https://github.com/ovh/manager/commit/a54a891caaa1c2e4042a48e974f0f0acec7f4f80))
+* **web:** dns history ([#10386](https://github.com/ovh/manager/issues/10386)) ([fde3865](https://github.com/ovh/manager/commit/fde386576225e29f1f6d9b84cf256d03c0af7cfc))
+
+
+
+
+
+## [18.8.2](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@18.8.1...@ovh-ux/manager-web@18.8.2) (2023-12-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-web
+
+
+
+
+
+## [18.8.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@18.8.0...@ovh-ux/manager-web@18.8.1) (2023-12-05)
+
+**Note:** Version bump only for package @ovh-ux/manager-web
+
+
+
+
+
 # [18.8.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@18.7.0...@ovh-ux/manager-web@18.8.0) (2023-12-04)