+++ packages/manager/modules/hub/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.14](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub@3.7.13...@ovh-ux/manager-hub@3.7.14) (2023-12-05)
+
+**Note:** Version bump only for package @ovh-ux/manager-hub
+
+
+
+
+
 ## [3.7.13](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub@3.7.12...@ovh-ux/manager-hub@3.7.13) (2023-12-01)
 
 **Note:** Version bump only for package @ovh-ux/manager-hub