+++ packages/manager/modules/mfa-enrollment/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.5](https://github.com/ovh/manager/compare/@ovh-ux/mfa-enrollment@2.4.4...@ovh-ux/mfa-enrollment@2.4.5) (2023-12-05)
+
+**Note:** Version bump only for package @ovh-ux/mfa-enrollment
+
+
+
+
+
 ## [2.4.4](https://github.com/ovh/manager/compare/@ovh-ux/mfa-enrollment@2.4.3...@ovh-ux/mfa-enrollment@2.4.4) (2023-11-21)
 
 **Note:** Version bump only for package @ovh-ux/mfa-enrollment