+++ packages/manager/modules/telecom-styles/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.6.8](https://github.com/ovh/manager/compare/@ovh-ux/manager-telecom-styles@4.6.7...@ovh-ux/manager-telecom-styles@4.6.8) (2023-12-05)
+
+**Note:** Version bump only for package @ovh-ux/manager-telecom-styles
+
+
+
+
+
 ## [4.6.7](https://github.com/ovh/manager/compare/@ovh-ux/manager-telecom-styles@4.6.6...@ovh-ux/manager-telecom-styles@4.6.7) (2023-12-01)
 
 **Note:** Version bump only for package @ovh-ux/manager-telecom-styles